requirements.txt
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pydantic==2.5.3
pyyaml==6.0.1
pytz==2023.3.post1
requests==2.31.0
python-dotenv==1.0.0
